Update Models/ - Suppression du code en double
