+++ src/main/java/models/BankItemTransactionSchema.java
@@ -1,18 +1,9 @@
 package models;
 
-public class BankItemTransactionSchema {
-    private CooldownSchema cooldown;
+public class BankItemTransactionSchema extends CommunSchema {
     private ItemSchema item;
     private SimpleItemSchema[] bank;
-    private CharacterSchema character;
 
-    public CooldownSchema getCooldown() {
-        return cooldown;
-    }
-
-    public void setCooldown(CooldownSchema cooldown) {
-        this.cooldown = cooldown;
-    }
 
     public ItemSchema getItem() {
         return item;
@@ -28,14 +19,6 @@ public class BankItemTransactionSchema {
 
     public void setBank(SimpleItemSchema[] bank) {
         this.bank = bank;
-    }
-
-    public CharacterSchema getCharacter() {
-        return character;
-    }
-
-    public void setCharacter(CharacterSchema character) {
-        this.character = character;
     }
 
 }
