+++ src/main/java/models/CharacterFightDataSchema.java
@@ -1,17 +1,7 @@
 package models;
 
-public class CharacterFightDataSchema {
-    private CooldownSchema cooldown;
+public class CharacterFightDataSchema extends CommunSchema {
     private FightSchema fight;
-    private CharacterSchema character;
-
-    public CooldownSchema getCooldown() {
-        return cooldown;
-    }
-
-    public void setCooldown(CooldownSchema cooldown) {
-        this.cooldown = cooldown;
-    }
 
     public FightSchema getFight() {
         return fight;
@@ -19,14 +9,6 @@ public class CharacterFightDataSchema {
 
     public void setFight(FightSchema fight) {
         this.fight = fight;
-    }
-
-    public CharacterSchema getCharacter() {
-        return character;
-    }
-
-    public void setCharacter(CharacterSchema character) {
-        this.character = character;
     }
 
 }
