+++ src/main/java/models/CharacterMovementDataSchema.java
@@ -1,17 +1,7 @@
 package models;
 
-public class CharacterMovementDataSchema {
-    private CooldownSchema cooldown;
+public class CharacterMovementDataSchema extends CommunSchema {
     private MapSchema destination;
-    private CharacterSchema character;
-
-    public CooldownSchema getCooldown() {
-        return cooldown;
-    }
-
-    public void setCooldown(CooldownSchema cooldown) {
-        this.cooldown = cooldown;
-    }
 
     public MapSchema getDestination() {
         return destination;
@@ -21,12 +11,5 @@ public class CharacterMovementDataSchema {
         this.destination = destination;
     }
 
-    public CharacterSchema getCharacter() {
-        return character;
-    }
-
-    public void setCharacter(CharacterSchema character) {
-        this.character = character;
-    }
 
 }
