+++ src/main/java/models/CommunSchema.java
@@ -1,0 +1,22 @@
+package models;
+
+public class CommunSchema {
+    private CooldownSchema cooldown;
+    private CharacterSchema character;
+
+    public CooldownSchema getCooldown() {
+        return cooldown;
+    }
+
+    public void setCooldown(CooldownSchema cooldown) {
+        this.cooldown = cooldown;
+    }
+
+    public CharacterSchema getCharacter() {
+        return character;
+    }
+
+    public void setCharacter(CharacterSchema character) {
+        this.character = character;
+    }
+}
