+++ src/main/java/models/DeleteItemSchema.java
@@ -1,17 +1,7 @@
 package models;
 
-public class DeleteItemSchema {
-    private CooldownSchema cooldown;
+public class DeleteItemSchema extends CommunSchema {
     private SimpleItemSchema item;
-    private CharacterSchema character;
-
-    public CooldownSchema getCooldown() {
-        return cooldown;
-    }
-
-    public void setCooldown(CooldownSchema cooldown) {
-        this.cooldown = cooldown;
-    }
 
     public SimpleItemSchema getItem() {
         return item;
@@ -19,13 +9,5 @@ public class DeleteItemSchema {
 
     public void setItem(SimpleItemSchema item) {
         this.item = item;
-    }
-
-    public CharacterSchema getCharacter() {
-        return character;
-    }
-
-    public void setCharacter(CharacterSchema character) {
-        this.character = character;
     }
 }
