+++ src/main/java/models/GETransactionListSchema.java
@@ -1,17 +1,7 @@
 package models;
 
-public class GETransactionListSchema {
-    private CooldownSchema cooldown;
+public class GETransactionListSchema extends CommunSchema {
     private GETransactionSchema transaction;
-    private CharacterSchema character;
-
-    public CooldownSchema getCooldown() {
-        return cooldown;
-    }
-
-    public void setCooldown(CooldownSchema cooldown) {
-        this.cooldown = cooldown;
-    }
 
     public GETransactionSchema getTransaction() {
         return transaction;
@@ -19,13 +9,5 @@ public class GETransactionListSchema {
 
     public void setTransaction(GETransactionSchema transaction) {
         this.transaction = transaction;
-    }
-
-    public CharacterSchema getCharacter() {
-        return character;
-    }
-
-    public void setCharacter(CharacterSchema character) {
-        this.character = character;
     }
 }
