+++ src/main/java/models/RecyclingDataSchema.java
@@ -1,17 +1,7 @@
 package models;
 
-public class RecyclingDataSchema {
-    private CooldownSchema cooldown;
+public class RecyclingDataSchema extends CommunSchema {
     private RecyclingItemsSchema details;
-    private CharacterSchema character;
-
-    public CooldownSchema getCooldown() {
-        return cooldown;
-    }
-
-    public void setCooldown(CooldownSchema cooldown) {
-        this.cooldown = cooldown;
-    }
 
     public RecyclingItemsSchema getDetails() {
         return details;
@@ -19,13 +9,5 @@ public class RecyclingDataSchema {
 
     public void setDetails(RecyclingItemsSchema details) {
         this.details = details;
-    }
-
-    public CharacterSchema getCharacter() {
-        return character;
-    }
-
-    public void setCharacter(CharacterSchema character) {
-        this.character = character;
     }
 }
