+++ src/main/java/models/SkillDataSchema.java
@@ -1,17 +1,7 @@
 package models;
 
-public class SkillDataSchema {
-    private CooldownSchema cooldown;
+public class SkillDataSchema extends CommunSchema {
     private SkillInfoSchema details;
-    private CharacterSchema character;
-
-    public CooldownSchema getCooldown() {
-        return cooldown;
-    }
-
-    public void setCooldown(CooldownSchema cooldown) {
-        this.cooldown = cooldown;
-    }
 
     public SkillInfoSchema getDetails() {
         return details;
@@ -19,14 +9,6 @@ public class SkillDataSchema {
 
     public void setDetails(SkillInfoSchema details) {
         this.details = details;
-    }
-
-    public CharacterSchema getCharacter() {
-        return character;
-    }
-
-    public void setCharacter(CharacterSchema character) {
-        this.character = character;
     }
 
 }
