+++ src/main/java/models/TaskCancelledSchema.java
@@ -1,22 +1,4 @@
 package models;
 
-public class TaskCancelledSchema {
-    private CooldownSchema cooldown;
-    private CharacterSchema character;
-
-    public CooldownSchema getCooldown() {
-        return cooldown;
-    }
-
-    public void setCooldown(CooldownSchema cooldown) {
-        this.cooldown = cooldown;
-    }
-
-    public CharacterSchema getCharacter() {
-        return character;
-    }
-
-    public void setCharacter(CharacterSchema character) {
-        this.character = character;
-    }
+public class TaskCancelledSchema extends CommunSchema {
 }
