+++ src/main/java/models/TaskDataSchema.java
@@ -1,17 +1,7 @@
 package models;
 
-public class TaskDataSchema {
-    private CooldownSchema cooldown;
+public class TaskDataSchema extends CommunSchema {
     private TaskSchema task;
-    private CharacterSchema character;
-
-    public CooldownSchema getCooldown() {
-        return cooldown;
-    }
-
-    public void setCooldown(CooldownSchema cooldown) {
-        this.cooldown = cooldown;
-    }
 
     public TaskSchema getTask() {
         return task;
@@ -19,13 +9,5 @@ public class TaskDataSchema {
 
     public void setTask(TaskSchema task) {
         this.task = task;
-    }
-
-    public CharacterSchema getCharacter() {
-        return character;
-    }
-
-    public void setCharacter(CharacterSchema character) {
-        this.character = character;
     }
 }
