+++ src/main/java/models/TaskRewardDataSchema.java
@@ -1,17 +1,7 @@
 package models;
 
-public class TaskRewardDataSchema {
-    private CooldownSchema cooldown;
+public class TaskRewardDataSchema extends CommunSchema {
     private TaskRewardSchema reward;
-    private CharacterSchema character;
-
-    public CooldownSchema getCooldown() {
-        return cooldown;
-    }
-
-    public void setCooldown(CooldownSchema cooldown) {
-        this.cooldown = cooldown;
-    }
 
     public TaskRewardSchema getReward() {
         return reward;
@@ -20,12 +10,5 @@ public class TaskRewardDataSchema {
     public void setReward(TaskRewardSchema reward) {
         this.reward = reward;
     }
-
-    public CharacterSchema getCharacter() {
-        return character;
-    }
-
-    public void setCharacter(CharacterSchema character) {
-        this.character = character;
-    }
+    
 }
